Legens block reorder in columns options.
When this option inside legends block it seted to true and it's neccesary more than one column for the legends block, a new algorithm computes the best distribution of the legend items inside the columns.

When a legend item is rendered to another column, it's brake legend name and legend icon. To fix it, please enable the dontBreakItems option.